requirements-datasci.txt
beautifulsoup4
notebook
SQLAlchemy
mysqlclient
geopy
pillow
psycopg
prompt-toolkit
ghapi
python-dotenv
anthropic
